webkit/data/layout_tests/chrome/fast/forms/textarea-metrics-expected.txt

   
This test checks that textareas have the right metrics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('textarea-no-scroll').clientWidth is 94
PASS document.getElementById('textarea-no-scroll').clientHeight is 94
PASS document.getElementById('textarea-no-scroll').offsetWidth is 96
PASS document.getElementById('textarea-no-scroll').offsetHeight is 96
PASS document.getElementById('textarea-no-scroll').scrollWidth is 94
FAIL document.getElementById('textarea-no-scroll').scrollHeight should be 20. Was 19.
FAIL document.getElementById('textarea-with-scroll').clientWidth should be 77. Was 79.
PASS document.getElementById('textarea-with-scroll').clientHeight is 94
PASS document.getElementById('textarea-with-scroll').offsetWidth is 96
PASS document.getElementById('textarea-with-scroll').offsetHeight is 96
FAIL document.getElementById('textarea-with-scroll').scrollWidth should be 77. Was 79.
FAIL document.getElementById('textarea-with-scroll').scrollHeight should be 132. Was 124.
PASS document.getElementById('textarea-no-scroll-with-padding').clientWidth is 106
PASS document.getElementById('textarea-no-scroll-with-padding').clientHeight is 106
PASS document.getElementById('textarea-no-scroll-with-padding').offsetWidth is 108
PASS document.getElementById('textarea-no-scroll-with-padding').offsetHeight is 108
PASS document.getElementById('textarea-no-scroll-with-padding').scrollWidth is 106
FAIL document.getElementById('textarea-no-scroll-with-padding').scrollHeight should be 32. Was 31.
FAIL document.getElementById('textarea-with-scroll-with-padding').clientWidth should be 89. Was 91.
PASS document.getElementById('textarea-with-scroll-with-padding').clientHeight is 106
PASS document.getElementById('textarea-with-scroll-with-padding').offsetWidth is 108
PASS document.getElementById('textarea-with-scroll-with-padding').offsetHeight is 108
FAIL document.getElementById('textarea-with-scroll-with-padding').scrollWidth should be 89. Was 91.
FAIL document.getElementById('textarea-with-scroll-with-padding').scrollHeight should be 144. Was 136.
